+++ sdk/kubernetesconfiguration/arm-kubernetesconfiguration-extensiontypes/src/api/extensionTypes/index.ts
@@ -1,0 +1,23 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export {
+  clusterListVersions,
+  clusterGetVersion,
+  listVersions,
+  getVersion,
+  list,
+  get,
+  locationList,
+  locationGet,
+} from "./operations.js";
+export type {
+  ExtensionTypesClusterListVersionsOptionalParams,
+  ExtensionTypesClusterGetVersionOptionalParams,
+  ExtensionTypesListVersionsOptionalParams,
+  ExtensionTypesGetVersionOptionalParams,
+  ExtensionTypesListOptionalParams,
+  ExtensionTypesGetOptionalParams,
+  ExtensionTypesLocationListOptionalParams,
+  ExtensionTypesLocationGetOptionalParams,
+} from "./options.js";