+++ sdk/kubernetesconfiguration/arm-kubernetesconfiguration-extensiontypes/CHANGELOG.md
@@ -1,4 +1,32 @@
 # Release History
+
+## 1.0.0-beta.2 (2026-03-13)
+Compared with version 1.0.0-beta.1
+
+### Features Added
+  - Added operation ExtensionTypesOperations.clusterListVersions
+  - Added operation ExtensionTypesOperations.locationList
+  - Added Interface ClusterScopeSettingsProperties
+  - Added Interface PagedAsyncIterableIterator
+  - Added Interface PageSettings
+  - Added Interface SystemData
+  - Interface ClusterScopeSettings has a new optional parameter systemData
+  - Interface ExtensionType has a new optional parameter systemData
+  - Interface ExtensionTypeVersionForReleaseTrain has a new optional parameter systemData
+  - Interface ProxyResource has a new optional parameter systemData
+  - Interface Resource has a new optional parameter systemData
+  - Added Type Alias AzureSupportedClouds
+  - Added Type Alias CreatedByType
+  - Added Enum AzureClouds
+  - Added Enum KnownCreatedByType
+  - Added Enum KnownVersions
+
+### Breaking Changes
+  - Removed operation ExtensionTypes.listClusterListVersions
+  - Removed operation ExtensionTypes.listLocationList
+  - Removed Interface ExtensionTypesList
+  - Removed Interface ExtensionTypeVersionsList
+
     
 ## 1.0.0-beta.1 (2025-05-19)